backend/pyproject.toml
[project]
name = "dbfinal-backend"
version = "0.1.0"
description = "Django REST API backend"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.6.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
    "django-filter>=24.0",
]
